Updated test to use the proper value
Do not change the value of maxVel you are making the test match the broken code.

The unit test overrides maxVel with this line  Constants.kPathFollowingMaxVel = maxVel;

When you debug the swerve module you can see the constant is set to 100 so my guess is that some piece of code is using a different constant that happens to be set to the same value of our yaml,  This is a bug waiting to happen when we change a value and something does not react.